+++ src/components/news-stand-latest/news-stand-rolling.js
@@ -1,4 +1,4 @@
-import { HEADER_CLASS, ROLLING, SIDE } from '../../constants/news-stand-rolling.js';
+import { HEADER_CLASS, ROLLING, SIDE } from '../../constants/news-stand-latest/news-stand-latest.js';
 
 class RollingNews {
   constructor(newsWrappers, intervals, side) {
