+++ src/constants/news-stand-header/news-stand-header.js
