+++ src/styles/components/news-stand-header/news-stand-header.css
