+++ src/styles/components/news-stand-latest/news-stand-latest.css
